refactor: extract chat logging functionality from `ChatLogger` to a new `ChatLogManager`.
